# gen-go/models/get_table_latency_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// GetTableLatencyResponse get table latency response
// swagger:model GetTableLatencyResponse
type GetTableLatencyResponse struct {

	// database
	// Required: true
	Database AnalyticsDatabase `json:"database"`

	// Table latency in hours
	Latency *float64 `json:"latency,omitempty"`

	// owner
	// Required: true
	Owner *string `json:"owner"`

	// schema
	// Required: true
	Schema *string `json:"schema"`

	// table
	// Required: true
	Table *string `json:"table"`

	// thresholds
	// Required: true
	Thresholds *Thresholds `json:"thresholds"`
}

// Validate validates this get table latency response
func (m *GetTableLatencyResponse) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateDatabase(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateOwner(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateSchema(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateTable(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateThresholds(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *GetTableLatencyResponse) validateDatabase(formats strfmt.Registry) error {

	if err := m.Database.Validate(formats); err != nil {
		if ve, ok := err.(*errors.Validation); ok {
			return ve.ValidateName("database")
		}
		return err
	}

	return nil
}

func (m *GetTableLatencyResponse) validateOwner(formats strfmt.Registry) error {

	if err := validate.Required("owner", "body", m.Owner); err != nil {
		return err
	}

	return nil
}

func (m *GetTableLatencyResponse) validateSchema(formats strfmt.Registry) error {

	if err := validate.Required("schema", "body", m.Schema); err != nil {
		return err
	}

	return nil
}

func (m *GetTableLatencyResponse) validateTable(formats strfmt.Registry) error {

	if err := validate.Required("table", "body", m.Table); err != nil {
		return err
	}

	return nil
}

func (m *GetTableLatencyResponse) validateThresholds(formats strfmt.Registry) error {

	if err := validate.Required("thresholds", "body", m.Thresholds); err != nil {
		return err
	}

	if m.Thresholds != nil {

		if err := m.Thresholds.Validate(formats); err != nil {
			if ve, ok := err.(*errors.Validation); ok {
				return ve.ValidateName("thresholds")
			}
			return err
		}
	}

	return nil
}

// MarshalBinary interface implementation
func (m *GetTableLatencyResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *GetTableLatencyResponse) UnmarshalBinary(b []byte) error {
	var res GetTableLatencyResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}